Move tests that require reinit and only run on .NET Framework
